Split test file per module
